fixup: refactor: promote CountHashDifficultyBits() to pkg/crypto/protocol
